Correcting: structure, naming

+++ src/main/java/com/example/testtask/api/dto/MeterReadingsDTO.java
@@ -6,6 +6,7 @@ import jakarta.validation.constraints.NotNull;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.time.LocalDateTime;
 import java.time.Month;
@@ -20,14 +21,14 @@ import java.time.ZoneId;
 public class MeterReadingsDTO {
 
     @NotNull(message = "ID не должен быть пустым")
-    long id;
+    Long id;
 
     @NotNull(message = "Значение не должно быть пустым")
     Month readingsDate;
 
     @NotNull(message = "Значение не должно быть пустым")
     @Digits(integer = 10, fraction = 3, message = "Значение должно быть в формате 0.001")
-    double readings;
+    BigDecimal readings;
 
     @NonNull
     @JsonProperty("creation_date")
@@ -39,3 +40,4 @@ public class MeterReadingsDTO {
     Instant updatedAt = Instant.now();
 
 }
+
